update opset and status

+++ onnx_tf/handlers/backend/sequence_construct.py
@@ -3,6 +3,7 @@ import tensorflow as tf
 from onnx_tf.handlers.backend_handler import BackendHandler
 from onnx_tf.handlers.handler import onnx_op
 
+
 @onnx_op("SequenceConstruct")
 class SequenceConstruct(BackendHandler):
 
@@ -11,7 +12,7 @@ class SequenceConstruct(BackendHandler):
     # create an empty sequence first
     tensor_dict = kwargs["tensor_dict"]
     dtype = tensor_dict[node.inputs[0]].dtype
-    input_sequence = tf.ragged.constant([], dtype = dtype)
+    input_sequence = tf.ragged.constant([], dtype=dtype)
 
     # insert tensors at the end of sequence
     for i in range(len(node.inputs)):
@@ -19,7 +20,7 @@ class SequenceConstruct(BackendHandler):
       if input_sequence.shape[0] == 0:
         output_seq = tf.RaggedTensor.from_tensor(input_tensor)
       else:
-        output_seq = tf.concat([input_sequence, input_tensor], axis = 0)
+        output_seq = tf.concat([input_sequence, input_tensor], axis=0)
       input_sequence = output_seq
 
     return [output_seq]
